Remove Action::FetchExistingMsgs
