+++ src/context.rs
@@ -434,6 +434,12 @@ impl Context {
                 .to_string(),
         );
         res.insert(
+            "fetched_existing_msgs",
+            self.get_config_bool(Config::FetchedExistingMsgs)
+                .await?
+                .to_string(),
+        );
+        res.insert(
             "show_emails",
             self.get_config_int(Config::ShowEmails).await?.to_string(),
         );
